Separate tests out by role
