+++ tests/Feature/Admin/Titles/RestoreTitleSuccessConditionsTest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Tests\Feature\Admin\Titles;
+
+use Tests\TestCase;
+use App\Models\Title;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group admins
+ */
+class RestoreTitleSuccessConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /** @test */
+    public function an_administrator_can_restore_a_deleted_title()
+    {
+        $this->actAs('administrator');
+        $title = factory(Title::class)->create(['deleted_at' => now()->toDateTimeString()]);
+
+        $response = $this->put(route('titles.restore', $title));
+
+        $response->assertRedirect(route('titles.index'));
+        $this->assertNull($title->fresh()->deleted_at);
+    }
+}
