+++ tests/Feature/Admin/Titles/UpdateTitleSuccessConditionsTest.php
@@ -1,0 +1,57 @@
+<?php
+
+namespace Tests\Feature\Admin\Titles;
+
+use Tests\TestCase;
+use App\Models\Title;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group admins
+ */
+class UpdateTitleSuccessConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /**
+     * Valid parameters for request.
+     *
+     * @param  array $overrides
+     * @return array
+     */
+    private function validParams($overrides = [])
+    {
+        return array_replace([
+            'name' => 'Example Name Title',
+            'introduced_at' => now()->toDateTimeString(),
+        ], $overrides);
+    }
+
+    /** @test */
+    public function an_administrator_can_view_the_form_for_editing_a_title()
+    {
+        $this->actAs('administrator');
+        $title = factory(Title::class)->create();
+
+        $response = $this->get(route('titles.edit', $title));
+
+        $response->assertViewIs('titles.edit');
+        $this->assertTrue($response->data('title')->is($title));
+    }
+
+    /** @test */
+    public function an_administrator_can_update_a_title()
+    {
+        $this->actAs('administrator');
+        $title = factory(Title::class)->create();
+
+        $response = $this->from(route('titles.edit', $title))
+                        ->patch(route('titles.update', $title), $this->validParams());
+
+        $response->assertRedirect(route('titles.index'));
+        tap($title->fresh(), function ($title) {
+            $this->assertEquals('Example Name Title', $title->name);
+        });
+    }
+}
