+++ tests/Feature/Guest/Titles/CreateTitleFailureConditionsTest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Tests\Feature\Guest\Titles;
+
+use Tests\TestCase;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group guests
+ */
+class CreateTitleFailureConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /**
+     * Valid parameters for request.
+     *
+     * @param  array $overrides
+     * @return array
+     */
+    private function validParams($overrides = [])
+    {
+        return array_replace([
+            'name' => 'Example Name Title',
+            'introduced_at' => today()->toDateTimeString(),
+        ], $overrides);
+    }
+
+    /** @test */
+    public function a_guest_cannot_view_the_form_for_creating_a_title()
+    {
+        $response = $this->get(route('titles.create'));
+
+        $response->assertRedirect(route('login'));
+    }
+
+    /** @test */
+    public function a_guest_cannot_create_a_title()
+    {
+        $response = $this->post(route('titles.store'), $this->validParams());
+
+        $response->assertRedirect(route('login'));
+    }
+}
