+++ tests/Feature/Guest/Titles/UpdateTitleFailureConditionsTest.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Tests\Feature\Guest\Titles;
+
+use Tests\TestCase;
+use App\Models\Title;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group guests
+ */
+class UpdateTitleFailureConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /**
+     * Valid parameters for request.
+     *
+     * @param  array $overrides
+     * @return array
+     */
+    private function validParams($overrides = [])
+    {
+        return array_replace([
+            'name' => 'Example Name Title',
+            'introduced_at' => now()->toDateTimeString(),
+        ], $overrides);
+    }
+
+    /** @test */
+    public function a_guest_cannot_view_the_form_for_editing_a_title()
+    {
+        $title = factory(Title::class)->create();
+
+        $response = $this->get(route('titles.edit', $title));
+
+        $response->assertRedirect(route('login'));
+    }
+
+    /** @test */
+    public function a_guest_cannot_update_a_title()
+    {
+        $title = factory(Title::class)->create();
+
+        $response = $this->patch(route('titles.update', $title), $this->validParams());
+
+        $response->assertRedirect(route('login'));
+    }
+}
