+++ tests/Feature/User/Titles/CreateTitleFailureConditionsTest.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Tests\Feature\User\Titles;
+
+use Carbon\Carbon;
+use Tests\TestCase;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group users
+ */
+class CreateTitleTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /**
+     * Valid parameters for request.
+     *
+     * @param  array $overrides
+     * @return array
+     */
+    private function validParams($overrides = [])
+    {
+        return array_replace([
+            'name' => 'Example Name Title',
+            'introduced_at' => now()->toDateTimeString(),
+        ], $overrides);
+    }
+
+    /** @test */
+    public function a_basic_user_cannot_view_the_form_for_creating_a_title()
+    {
+        $this->actAs('basic-user');
+
+        $response = $this->get(route('titles.create'));
+
+        $response->assertForbidden();
+    }
+
+    /** @test */
+    public function a_basic_user_cannot_create_a_title()
+    {
+        $this->actAs('basic-user');
+
+        $response = $this->post(route('titles.store'), $this->validParams());
+
+        $response->assertForbidden();
+    }
+}
