+++ tests/Feature/User/Titles/UpdateTitleFailureConditionsTest.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Tests\Feature\User\Titles;
+
+use Tests\TestCase;
+use App\Models\Title;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group users
+ */
+class UpdateTitleFailureConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /**
+     * Valid parameters for request.
+     *
+     * @param  array $overrides
+     * @return array
+     */
+    private function validParams($overrides = [])
+    {
+        return array_replace([
+            'name' => 'Example Name Title',
+            'introduced_at' => now()->toDateTimeString(),
+        ], $overrides);
+    }
+
+    /** @test */
+    public function a_basic_user_cannot_view_the_form_for_editing_a_title()
+    {
+        $this->actAs('basic-user');
+        $title = factory(Title::class)->create();
+
+        $response = $this->get(route('titles.edit', $title));
+
+        $response->assertForbidden();
+    }
+
+    /** @test */
+    public function a_basic_user_cannot_update_a_title()
+    {
+        $this->actAs('basic-user');
+        $title = factory(Title::class)->create();
+
+        $response = $this->patch(route('titles.update', $title), $this->validParams());
+
+        $response->assertForbidden();
+    }
+}
