+++ tests/Feature/User/Titles/ViewTitlePageFailureConditionsTest.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tests\Feature\User\Titles;
+
+use Tests\TestCase;
+use App\Models\Title;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+
+/**
+ * @group titles
+ * @group users
+ */
+class ViewTitlePageFailureConditionsTest extends TestCase
+{
+    use RefreshDatabase;
+
+    /** @test */
+    public function a_basic_user_can_view_a_title()
+    {
+        $this->actAs('basic-user');
+        $title = factory(Title::class)->create();
+
+        $response = $this->get(route('titles.show', ['title' => $title]));
+
+        $response->assertForbidden();
+    }
+}
